+++ core/src/main/java/io/snabble/sdk/CustomerCardInfo.java
@@ -1,0 +1,19 @@
+package io.snabble.sdk;
+
+public class CustomerCardInfo {
+    private String cardId;
+    private boolean isRequired;
+
+    CustomerCardInfo(String cardId, boolean isRequired) {
+        this.cardId = cardId;
+        this.isRequired = isRequired;
+    }
+
+    public String getCardId() {
+        return cardId;
+    }
+
+    public boolean isRequired() {
+        return isRequired;
+    }
+}